+++ src/utils/image-src/generate-images-path.js
@@ -2,7 +2,7 @@ import { FALSY_VALUES } from '../../constants/falsy-values';
 import { getResponsiveWidthOfContainer } from '../responsive/get-responsive-width-of-container';
 import { getSizeAccordingToPixelRatio } from '../responsive/get-size-according-to-pixel-ratio';
 
-export const generateImagesPath = (srcConfig) => {
+export const generateImagesPath = (srcConfig, loadOriginalImages) => {
   const {
     container, folder, apiVersion, filename = '', ciParams,
   } = srcConfig;
@@ -23,7 +23,7 @@ export const generateImagesPath = (srcConfig) => {
     const cdn = isCloudImageUrl ? src
       : `https://${ciToken}.cloudimg.io/${finalApiVersion}${src}`;
 
-    src = `${cdn}?${ciTransformation || `width=${ciSizeNext}`}${ciFilters ? `&f=${ciFilters}` : ''}`;
+    src = `${cdn}?${ciTransformation || `${!loadOriginalImages ? `width=${ciSizeNext}`: ''} `}${ciFilters ? `&f=${ciFilters}` : ''}`;
   }
 
   return src;